refactor(footer): supprime l'ancien composant Contact et déplace l'actuel
